pyproject.toml
[tool.poetry]
name = "dogeaction"
version = "0.1.0"
description = ""
authors = ["Carlos Gonzalez <carlos.gonzalez@with-madrid.com>"]

packages = [
    { include = "dogeaction" },
]

[tool.poetry.dependencies]
python = "^3.9"
typefit = "^0.4.0"
actions-toolkit = "^0.1.13"
PyYAML = "^6.0"
typer = {extras = ["all"], version = "^0.7.0"}

[tool.poetry.group.dev.dependencies]
ipython = "^8.11.0"
isort = "^5.10.1"
black = "^22.3.0"
pytest = "^7.1.2"
gitpython = "^3.1.31"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
doge = "dogeaction.app:app"
